begin working on converting startup scripts to a proper package
