+++ startup/14-move_mirrors.save
@@ -2,7 +2,7 @@
 import bluesky.plan_stubs as bps
 from numpy import tan, pi
 
-run_report(__file__)
+run_report(__file__, text='mirror trigonometry')
 
 
 def move_m3(target=5):
