+++ startup/42-fe-bpm.py
@@ -1,10 +1,8 @@
-from ophyd import (SingleTrigger, Component as Cpt, Device, DeviceStatus, EpicsSignalRO)
+
+from BMM.frontend import FEBPM
 
 run_report(__file__)
 
-class FEBPM(Device):
-    x = Cpt(EpicsSignalRO, 'X-I')
-    y = Cpt(EpicsSignalRO, 'Y-I')
 
 try:
     bpm_upstream   = FEBPM('SR:C06-BI{BPM:4}Pos:', name='bpm_upstream')
