+++ startup/94-base.py
@@ -1,0 +1,8 @@
+
+
+run_report(__file__, text='final chores')
+
+from BMM.motor_status import motor_metadata, motor_status, ms, motor_sidebar, xrd_motors, xrdm
+
+from BMM.linescans import linescan
+
